Improve error handling for API status 404
- Refactor weather information into WeatherInfo component
- MessageFlash component no longer used for 404 error messages
- If API returns a 404 status code, display an error message and
  remove existing weather information
