+++ src/stylesheets/MessageFlash.css
@@ -3,10 +3,6 @@
   min-height: 1.2rem;
 }
 
-.error {
-  background-color: #ffcccc;
-}
-
 .tool-tip {
   background-color: #eeeeee;
 }
